new brushes: characoal, airbrush...
The airbrush "real" try to emulate the movement of the hand in one real
airbrush. Slow for details and fast for big diemeter and with the
presion more o less paint. But is not necessary use the presion, you can
use "S" and "A" for start to paint, to acomplish a realistic feeling, like
if you are using a real airbrush. :) The digital airbrush is more easy to
use but not with a real movement of hand.