New brushes: Oils,Acrylics,Watercolors...
Oils:
The painting is not infinite, when  the brush is not painting anymore start
to smudge but with more presion you can paint a little.
Each paintbrush have two variants for use combined:
To paint(Paint).
Brush Clean(Clean)): it's equals to "Paint" but clean, only smudge the
paint.
and finally:
Mop: it's a paintbrush "mop" for smooth the paint.
---------------------------
Acrylics:
Each paintbrush have tree variants for use combined:
To paint(Paint).
To paint with more water (with water).
To use only water(only water).
--------------------------
Watercolor:
Each paintbrush have two variants for use combined:
To paint(Paint).
To use only water(only water).
---------------------------
Water:
they are for splash water, work without the presion, with the presion and
with the velocity, for differents effects.
---------------------------
Decoration: nothing special but experimenting with them I could make the
paintbrushes.
---------------------------
Texture: they are very utils for works of textures, and for speed sketches.