Differentiate release version and API version.
Cherry picked from commit 8feee81c6c849e419c4cf4502eaff1f1924e700c,
then adjusted for version 2.0. Currently the next release version will
be 2.0 as well. But later, it can be 2.1, 2.2, and so on. On the other
hand, the API version should stay 2.0. This is forward compatibility
indicating that the 2.0 brush format should not break (if it does, then
it will be time to go for 3.0 brushes).